database.py
import psycopg2

try:
    # пытаемся подключиться к базе данных
    conn = psycopg2.connect(dbname='database', user='postgres', password='postgres')
except:
    # в случае сбоя подключения будет выведено сообщение в STDOUT
    print('Can`t establish connection to database')