pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["georg <georbearwolf.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.8.1"
Flask = "2.2.2"
gunicorn = "20.1.0"
python-dotenv = "0.21.0"
poetry = "1.6.1"
psycopg2-binary = "^2.9.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
